Add files via upload
The files with prefix "nr" are for the "new rate" function data which is what is uploaded to BOX in the tar.bz2 file. The files with prefix "fs" are for "final state" because I was trying to figure out ways to predict the final states of the trajectories. Those files should work with all the other data on BOX.